DTO 매핑 리팩토링
- 기존 Service계층에 존재하던 DTO 매핑 메서드를 DTO 내부에 정적팩토리메소드로 변경(LineResponse, StationResponse)
