+++ src/main/java/subway/line/dto/LineResponse.java
@@ -1,8 +1,10 @@
 package subway.line.dto;
 
-import subway.station.StationResponse;
+import subway.line.Line;
+import subway.station.dto.StationResponse;
 
 import java.util.List;
+import java.util.stream.Collectors;
 
 public class LineResponse {
 
@@ -14,11 +16,18 @@ public class LineResponse {
 
     private List<StationResponse> stations;
 
-    public LineResponse(Long id, String name, String color, List<StationResponse> stations) {
+    private LineResponse(Long id, String name, String color, List<StationResponse> stations) {
         this.id = id;
         this.name = name;
         this.color = color;
         this.stations = stations;
+    }
+
+    public static LineResponse ofEntity(Line line) {
+        List<StationResponse> stations = line.getStations().stream()
+                .map(StationResponse::ofEntity).collect(Collectors.toList());
+
+        return new LineResponse(line.getId(), line.getName(), line.getColor(), stations);
     }
 
     public Long getId() {
