+++ src/main/java/subway/section/SectionRepository.java
@@ -1,9 +1,6 @@
 package subway.section;
 
 import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.data.jpa.repository.Query;
-
-import java.util.List;
 
 public interface SectionRepository extends JpaRepository<Section, Long> {
 }
