+++ src/main/java/subway/section/SectionService.java
@@ -2,10 +2,9 @@ package subway.section;
 
 import org.springframework.stereotype.Service;
 import subway.line.Line;
-import subway.line.exception.LineException;
 import subway.line.service.LineDataService;
 import subway.station.Station;
-import subway.station.StationDataService;
+import subway.station.service.StationDataService;
 
 import javax.transaction.Transactional;
