+++ src/main/java/subway/station/service/StationDataService.java
@@ -1,7 +1,9 @@
-package subway.station;
+package subway.station.service;
 
 import org.springframework.stereotype.Service;
 import subway.line.exception.LineException;
+import subway.station.Station;
+import subway.station.StationRepository;
 
 import javax.transaction.Transactional;
 
@@ -18,9 +20,5 @@ public class StationDataService {
     public Station findStation(Long id) {
         return stationRepository.findById(id)
                 .orElseThrow(() -> new LineException("존재하지 않는 역입니다."));
-    }
-
-    public StationResponse mappingToStationResponse(Station station) {
-        return new StationResponse(station.getId(), station.getName());
     }
 }
